Worked around a nasty issue in appending to $postgresql_auth_lines array
Previously the value of $postgresql_auth_lines was [] by default. If the array
contained only one entry after all external modules (e.g. bacula and trac) had
appended lines to the array, Puppet would convert it to a string before passing
it to the ERB template parser. This caused an ERB template parser error as a
string is not iterable. This bug was invisible if the array contained two or
more lines, and probably if it was empty.

Worked around this issue by changing the default value of $postgresql_auth_lines
into ['','']. While this is hackish, it seems to solve the problem at the cost
of adding two extra empty lines into pg_hba.conf.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>